async-pair: add DexFuture to wrap simple async/finish pairs
This doesn't handle anything without parameters or beyond the very simple
async/finish pairs, but you can cover a lot of ground with just that.

Using this you can bridge old-stlye GAsyncReadyCallback (and GTask)
oriented APIs into DexFuture for chainability.

Some work is still necessary so that you can propagate future cancellation
from the DexFuture to a GCancellable for these old APIs, but that will
come once we have a muxer future and discard propagation.