future-set: add ordered set of futures for result combining
The DexFutureSet future allows completion of a DexFuture (the DexFutureSet)
when a given number of dependent futures have either resolved or rejected.

This is currently used for dex_future_all() and dex_future_any(), but will
be used for more in the future.

You can retrieve the value for sub-futures using the provided
dex_future_set_get_size() and dex_future_set_get_future() functions.