future-set: fix const'ness of incoming DexFuture parameter
We want to be able to have these const, which means the inner * needs to
be const so that the outer array is const (not the DexFuture). The spiral
philosophy for reading types in C at work here.

Beyond that, we might as well take a ref in the callee so that we can have
a transfer notation that is representable by GObject Introspection as well.
It would be rather annoying to pass arrays and deal with proper free funcs
here when this API is mostly for bindings or situations where you have
allocations already.