block: save thread default scheduler for block execution
This just saves the scheduler instance so that we can use it to dispatch
the callback on the proper scheduler (still to be done). The goal here is
to ensure that things like "then" are only run in the same scheduler they
are created so that tooling like GTK which requires main thread access will
automatically do the right thing.