block: give access to DexBlock type in public API
It can be useful to be able to at least type check the DexFuture types
and get a bit of information about them since we have so many sub-types
available for DexFuture.

Additionally, pass the scheduler when creating the block so that we can
give access to what scheduler will be used.