thread-pool-scheduler: add scaffolding for scheduler
This just gets the subclass'ing in place for the threadpool scheduler. I
want to avoid using GThreadPool under the hood for a few reasons, and the
API not exposing a GThreadPool reflects that.

In particular, the locking mechanisms behind GThreadPool are ancient in
design. There is no work stealing, there are no integrated main contexts,
and there is no pinning of threads to CPUs.

The lack of pinning is probably alright, but ideally we'd get some
coordination to spin down our threads when CPUs go offline. If we have
a global queue for new work items not originating on the scheduler, that
would basically happen for free.

I want to have a main context that can integrate so we can have GSource-
based futures stay on the same thread as the block dispatching for then,
catch, finally, etc. Otherwise we bounce between the main thread and the
threadpool regularly which is not ideal.

Doing so, however, requires that we pump the threadpool work via a GSource
so the main loop itself can drive timeouts.

All of the above, basically makes the opaque'ness of GThreadPool unsuitable
for the backing mechanism of the scheduler.