thread-pool-worker: construct, finalization strategies
This still isn't complete as we haven't chosen our queue strategy for the
worker. I'd like to avoid lots of cross-thread frees, so some research is
still necessary to make a choice.

Additionally, we'd want the workers to be used as a set from the scheduler
and allow for rebalancing between them.