thread-storage: only take a single GPrivate entry
There are only 128 entries for GPrivate and so we want to avoid taking
many entries. This makes it so we only take a single entry (allowing the
rest for other libraries, including GLib), and pointer chases to the
table of state we care about from there.

We store the current DexThreadPoolWorker there so that we can use it in an
upcoming commit to push a work item onto the same worker instead of a
global work queue.