thread-pool-scheduler: push work item into current work queue
If a work item is pushed into the DexThreadPoolScheduler from a
DexThreadPoolWorker, then we want to enqueue that work item into the same
scheduler. This increases the chance that we will get a cacheline hit for
the data within the work item but also allows us to do it without heavy
synchronization operations across threads.

We still need to ensure the current scheduler is set in the target thread
so that this will get dispatched via the thread-default scheduler.