work-queue: start on a global work queue (MPMC)
The global work queue is for work items that are created outside of a
worker thread but associated with the DexThreadPoolScheduler.

Some research still needs to be done on what the best way to signal the
workers to wakeup if they're currently blocked in their GMainContext on
a poll().

Additionally, we need to figure out how to wakeup other workers when a
single worker has created a lot of backlog to churn through yet no new
work items have reached the global work queue.