work-stealing-queue: use aligned allocation for queue
If we are to embed the queue within the worker, we have the benefit that
we can avoid an additional pointer chase. However, that comes at the cost
of breaking the allocation requirements of GTypeInstance (which can only
guarantee an alignment of up to 8 bytes). Embedding our structure bumps
that requirement to 64 so that we don't have false sharing between various
atomics we're working with.

If we instead allocate the queue, and use aligned allocs for it, we can
maintain the alloc requirement from GTypeInstance at the cost of the extra
pointer chase (which is probably a non-issue anyway).