semaphore: start on FD-based semaphore design
This currently just uses eventfd() with EFD_SEMAPHORE. However, it's not
ideal yet in that you'll still have a thundering herd because we're using
poll() and not yet using edge-triggered interrupts.

Some work will need to be done to adapt those to work. We can probably make
the poll() side work with an indirection epollfd, but that still needs to
be tested.