uring: add scaffolding for aio with io_uring
This gets some bits in place for a GSource that manages an io_uring using
liburing. The goal here is to provide some minimal async I/O APIs that
will integrate with our other Future-based APIs.

Locally, I have something working that does a shared semaphore on io_uring,
so we should be able to eventually build on this for doing DexSemaphore
for the global work queue.