semaphore: make DexSemaphore work with io_uring
We will implement this differently in the future, but this at least shows
that it can work.

We'd probably implement the submission queue stuff on top of the
DexAioContext (which would use io_uring itself). It requires an extra
allocation for the future, but at least we wouldn't need a separate
io_uring just for this.