scheduler: incremental work on scheduler redesign
Now that we have more idea of how the schedulers will work and how async
IO/queues will connect with that, this starts progressing the
DexThreadPoolWorker to itself become a DexScheduler.

I still have a lot of unfinished parts in various corners, but I don't
like having so much not committed so we'll just keep iterating step-by-
step here.