main-scheduler: use DexWorkQueue for work items
This is a bit less performance oriented in that we're going through a FD
to notify our own thread, but it does allow for more testing/use of the
same work queue paths.

I'd generally like less duplication even if it means slightly slower right
now. We can optimize in the future if it makes sense.