work-queue: run queue from future loop instead of GSource
This allows us to re-use the DexAioContext instead of having that
duplicated within dex-semaphore.c. That is the first step to making
DexSemaphore supported on more platforms (it still currently relies on
eventfd and that needs fixing).

Additionally, make DexWorkQueue a DexObject so we can add more assertions
within the codebase to suppliement correctness checks.