build: remove generated io_uring version header
I submitted a PR upstream to io_uring to give us this version information
and hopefully we can just start requiring that at some point instead.

As it were, we aren't even using this yet, so just remove it while leaving
it available in the history to revert if we end up needing our own to
support older liburing versions.