channel: add DexChannel
This is used to communicate between two sides, somewhat like a pipe.
However, it allows a capacity to be specified so that only N items may
be in the primary queue as "accepted". The rest are queued in a special
queue until capacity is reduced.

This can be useful for scenarios where you want to communicate with a
peer, but you want to use the delayed nature of futures to automatically
reduce the velocity of the producer. If the producer limits itself by
awaiting the send future, it will only go up to the rate of the
capacity+1.

Order within the channel is preserved, even if items later in the channel
queue have completed and earlier items have not.