timeout: implement discard
Additionally, leave a note about how this returned from some functions
which then unref the future could cause the timer to never fire, as
nothing is holding a reference to it (and the source uses a weak ref).

This could be addressed by propagating await counts and discarding when
the await count reaches zero rather than by reference counts alone.