examples: add cat example
This can still be simplified a bunch, and I'd like to see that happen, but
I at least want to get something committed that works. We can also add a
variant of this eventually which uses DexAioBackend instead of the input
stream so that we can compare io_uring in the active scheduler against
GLocalFile using IO threads for the work (and bouncing between threads).