examples: simplify cat example
Make some shared code that we can port cat-aio to as well, and keep the
example a bit shorter. I still want to experiment with a DexFiber
eventually because we can make this code much easier to reason about,
even if we reduce it to limited cases where m:n threading might be
wanted.