fiber: test switching back to the original context
This is still using ucontext_t but we'll likely change that or use
something similar to libtask to abstract using those same types to avoid
the massive overhead of the OS's swapcontext().

Also, make stacks managed by the fiber implementation.

We might break this out into another source file so that we can do pooling
of stacks, but I want to at least hide the implementation from the public
API exposed to developers.