fiber: make DexFiber a DexFuture
This will allow awaiting on the completion of a fiber, which is much
more useful than our DexRoutine stuff (which I hope to eliminate after
this is finished).

There is still work to be done on propagation so that we can move the
fiber to the ready queue (and wakeup source if necessary). But I need a
break and so this is a decent pausing spot.