object: remove lots of get_type() function calls
Not only do those calls do an atomic_get() via g_once_init_enter(), but
they include a call for a type that is already locally available in the
module so long as the type is initialized in dex_init() properly. Since
we require either dex_init() to be called (if static linking) or for it
to be a static constructor (when dynamic linking), we can rely on the
dex_foo_type style addresses to already contain the GType. Just short-
circuit and use that.