fiber: import various libtask ucontext abstractions
We want to be able to support user-space switching on a number of platforms
where that isn't being done already. So this brings in the bits that are
relatively tested from Russ Cox's libtask, so that we have something to
start from.

This still leaves out Windows which we can eventually use the integrated
OS fiber APIs for what we need.