main-scheduler: use private work queue implementation
Instead of relying on DexWorkQueue, which itself relies on DexSemaphore,
this uses a private GSource implementation with a GQueue to process work
items for the main context.

Doing so means that we can actually test DexSemaphore using a proper
DexMainScheduler on that thread instead of the half-setup TLS data that
was being done previously. Otherwise, we are in a situation of semaphores
trying to use semaphores and nothing getting scheduled.