semaphore: implement fallback semaphore with futures
Ideally, we're using io_uring with eventfd() so that we can save a lot of
kernel cotext switches for the semaphore. However, in cases where we cannot
do that (such as macOS or Windows) we can do that with a DexFuture that is
completed through a work item scheduled on the thread-default scheduler.

This is a bit heavier than the eventfd()+io_uring case because we have to
wakeup N threads using a work item. Some other systems may do fairly well
if they have the combination of eventfd() and kqueue for IO (such as
FreeBSD 13+). Although, not a lot of testing has happened there yet beyond
just making some things compile/test.