block: restrict synchronous dispatch recursion
If we have a semaphore with a large number that completes synchronously,
then we could have considerable recursion beyond what we'd expect. To help
limit that, we can put in a max recursion depth so that we force ourselves
to allocate and skip back to processing work items on the next round of
the main context.