posix-aio-backend: add simple aio backend using a threadpool
This just defers the work to a threadpool, similar to what GIO does using
the various streams API. This is file-descriptor based of course so that
we can still take advantage of the various io_uring APIs in our plumbing,
but that could change depending on how low we want to push this all.

For example, if this were something we'd want inside of GLib, I'd probably
re-think the various AIO backends to fit within that context.