fiber: use allocation if alignof(ucontext_t) > sizeof(void*)
If we have an alignment requirement for ucontet_t that is larger than the
sizeof(void*) then we cannot embed the field in our struct as casts
would increase the alignment requirements.

Instead of loosing the optimization, try to accomidate such a situation
with an extra allocation while keeping the inline data for platforms
where that alignment requirement does not exist.