thread-pool-worker: short-circuit if aio context fails
io_uring will impose limits to us on the number of rings we can create,
so if we fail to create an io_uring it is a likely case that we've hit
a maximum limit. Since our thread pool is for dealing with Future-based
work (as opposed to genreal threadpool for long running, blocking tasks),
we can take that as a sign we should just stop creating more workers.