platform: handle sysconf(_SC_THREAD_STACK_MIN) failing gracefully
I don't know a good reason *why* it would fail; that is, unlike with PATH_MAX,
it seems always reasonable to enforce some minimum size. But in case it does
fail, do not mistake the returned value of -1 for a minimum of LONG_MAX.

Also, assert that the stack size we end up with, after all of our logic, looks
at least somewhat reasonable. Namely, it should not be less than a single page,
and it should not be more than 4 GB. These assertions would have caught the bug
that happened on the Hurd, where size ended up equal to zero (rounded up to
the page size from LONG_MAX), which manifested as a segfault inside a
makecontext () call down the road.