fiber-context: abstract ucontext_t usage
This creates a DexFiberContext type which is either a ucontext_t (when
alignment is <=sizeof(void*) or a ucontext_t (when the alignment requires
something >sizeof(void*).

The abstraction not only reduces the difficulty of reading dex-fiber.c,
but also allows us to abstract other fiber implementations. Notably, I
want to find a matching abstraction for the win32 fiber API.

This might break some systems, but most likely just because I haven't
tested the 32-bit compile or ucontext_t* builds yet (and will shortly).