fiber: fix swapcontext() on macOS
This broke because we started using 4 int parameters for spawning contexts
on x86_64. However, that breaks due to dex-ucontext.c which only supports
2 integers. Rather than try to make sure that we can support that without
trapping everywhere, it's easy enough to store the fiber hook in the figer
itself along with the ucontext_t.

This uses that address containing the function and backpointer to the fiber
to ensure that dex_fiber_start() is called correctly.

With this in place, tests pass on macOS again.