threadpool: clamp num threads to 32
If we end up on a huge n-way thread system, we want to avoid creating
hundreds of threads. This just clamps is to a lower value until we can
get around to controlling thread creation dynamically. That will take
additional work though so that we can steal from peers without having
to lock/sync to access the structure.