uring: limit io_uring support to Linux 6.1+
It appears that getting the support matrix right for 5.x kernels is quite
a bit of work and it might just be easier to set a low bar of a 6.x kernel
for things going forward. Given how few applications we have so far using
libdex, particularly for io_uring, this seems like a smart bet for longer
term support.

See #17 for more information.