thread-pool-worker: push() directly to a worker is high-priority
These need to be processed without going through an additional global
work queue as they tend to be the completion of DexFuture to DexBlock
similar.

We don't want those to get caught up in (re)queuing when the result is
already available.

If we want to make this perform better under heavy load, we can look to
create a single GSource per worker which processes these in batch without
the extra allocation and overhead of GSource/GMainContext.

Related to #17