semaphore: improve waiter completion w/o eventfd/liburing
We want to be a bit smarter in how we proceed without either eventfd or
liburing as well as without liburing but with eventfd.

This cleans up our counter usage and avoids using eventfd unless we also
have the liburing-based AIO back-end. Otherwise we just end up blocking
our IO threadpool with read() nonsense for the semaphore.

Related #17