scheduler: add G_WARN_UNUSED_RESULT to dex_scheduler_spawn()
This is a common thing to get a function into a new fiber. It is very easy
to forget to do something with the result of that fiber if you are just
doing processing work there.

To help people avoid leaking that DexFiber structure, add a warning for
compilers that support it.