future: keep discarded until after we unlock future
The goal here is to ensure that the WeakRef system which is used by the
ChainedFuture cannot in any way be related to the @future itself. While
that should be the case, this enforces it with minimal overhead since our
queue manipulation is O(1) * O(n) where n is number of chained futures.