build: respect meson's default_library option
Users may wish to embed / statically link to libdex e.g. as a subproject
fallback. libdex itself already supports linking tests to an internal
static library, but it shares the same name as the public interface so
it's not as simple as just switching `shared_library()` to `library()`.

We handle this with a bit of manual option unrolling. If
-Ddefault_library is set to shared, we need to build a separate static
library for internal use, which we call "dex-internal" to prevent
-uninstalled.pc usage picking it up inadvertently. In all other cases,
we grab the meson variable that represents the static part of the
`library()` build (for "both" it will have a dedicated function to
access this) and alias that to the existing variable.

Supporting default_library exposes an issue that the introspection gir
cannot be built using a static-only libdex. Raise an explicit error in
the case that introspection is also enabled.