thread-pool-scheduler: avoid extra pointer chase through pdata
We have a max number of workers so just allocate that directly in the
scheduler struct and avoid an extra pointer chase through pdata[].

This also ensures that the allocation underneath the workers array can
never change (so caching it is fine on any core). Though we aren't
giving our pointer to the scheduler out until after this function exits,
if a worker created a work item as part of startup it could have a
cached value for pdata right before the array expands.

Related: #142