block: don't dispatch blocks on fibers
If we're running within a fiber, defer the dispatch until the block has
been scheduled on the regular main loop scheduler. This provides more
reliable call sites because some blocks may not necessarily expect to
be run on a fiber.

This appears to fix an issue I'm seeing with rsvg/cairo not doing so
hot when getting initialized on a fiber.

There can of course be performance impacts with this because you need to
ping-pong back to the GMainContext scheduler, but it hasn't been too
problematic in what I've seen thus far.